Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -43,7 +43,7 @@ end
 def get_average_age_for_season(data, season)
   total = 0
   data[season].each do |contestant|
-    total += contestant["age"].to_i
+    total += contestant["age"].to_f
   end
   (total/ (data[season].length) )
 end
